% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iForm.R
\name{iformselect}
\alias{iformselect}
\title{Selection for iForm procedure}
\usage{
iformselect(x, y, p, n, C, S, bic, heredity, higher_order)
}
\arguments{
\item{x}{data.frame of your data with all p predictors}

\item{y}{vector of observed responses}

\item{p}{number of predictors in the dataset}

\item{n}{number of observations in the dataset}

\item{C}{vector of candidate predictors to consider in this step of the procedure}

\item{S}{vector of solution predictor selected from previous steps of the procedure}

\item{bic}{vector of bic values calculated for each step of the procedure}

\item{heredity}{a string specifying the heredity to be considered. NULL, weak, strong}

\item{higher_order}{logical TRUE indicating to include order-3 interactions in the search (default FALSE)}
}
\value{
the response vector, the solution set of predictors and the calculated bic values
}
\description{
This is a helper function to run the selection procedure under differen heredity principles
and different levels of interactions included in the selection.
}
\details{
Runs the iFORM selection procedure for specifed heredity and level of interactions.
It returns the solution to be fit from \code{iForm}
}
\author{
Kirk Gosik
}
