% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iForm.R
\name{rss_map_func}
\alias{rss_map_func}
\title{Finding minimum RSS}
\usage{
rss_map_func(C, S, y, data)
}
\arguments{
\item{C}{vector of candidate predictors to consider in this step of the procedure}

\item{S}{vector of solution predictor selected from previous steps of the procedure}

\item{y}{vector of observed responses}

\item{data}{data.frame of your data with the response and all p predictors}
}
\value{
A vector of the RSS values for each candidate predictor
}
\description{
Helper function to take in the candidate set and solution set along with the observations
and previous data to calculate the residual sum of sqaures for each of the candidate
predictors given what has already been selected.
}
\details{
Mapping function to calculcate the residual sum of squares for each of the candidate predictors
}
\author{
Kirk Gosik
}
