% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iForm.R
\name{strong_order3}
\alias{strong_order3}
\title{Creating interactions based off of strong heredity principle}
\usage{
strong_order3(S, data)
}
\arguments{
\item{S}{vector of solution predictor selected from previous steps of the procedure}

\item{data}{data.frame of your data with the response and all p predictors}
}
\value{
A vector of the RSS values for each candidate predictor
}
\description{
Helper function to give all possible order-3 interactions following the strong
heredity principle.
}
\details{
Finds all p choose 3 combinations between the predicotrs in the solution set.
}
\author{
Kirk Gosik
}
